Remove constituency filter from position pages
This doesn't work in it's current state so the simplest option is to
remove it. If people want to see the politicians for an area then they
can search for the area and go to the page for an individual area. These
pages then have a list of the representatives on them.