Remove css files before generating the site
Due to the way jekyll-compass works, the css doesn't get regenerated
when the repo is freshly cloned because it checks timestamps to see if
it needs to be regenerated.

To work around this we simply remove the css directory before running
'jekyll build', which forces the css to be recompiled.